Palya_valtoztatasok
+ Pálya elemek
+ Checkpoints
+ Trap Damage script / Trappek

+++ 2d_platformer_game/Assets/Scripts/Health.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public class Health : MonoBehaviour
 {
-    public static double totalHealth = 100;
+    public static float totalHealth = 100f;
 
 }
